Take home assessment
Hello and thanks for allowing me to take a stab at this assessment!

I wasn't sure what to expect and was pleased with the challenges given. There were a few aspects of the challenge that were new to me, and gave me an opportunity to try my best to come up with solutions.

First, I've never used Guzzle before, which is what I chose to go with for accessing the external API endpoint. Second, I've never used a Google Cloud function before - which provided me with the greatest challenge. Given the requirement to use an authentication header and a value associated with it, I assumed it was basic http authentication, but wasn't ever able to get the API to respond. So either I was either incorrect about how to authenticate for API access, or the API wasn't working for me as it returned an HTTP 403 status code. In addition to using the built in auth parameter with Guzzle, I also tried sending a custom header with basic authentication syntax in place, which returned me with a HTTP status code of 400. I did install and look into using Google's Cloud SDK in an attempt to try to get an OAuth token, however I didn't have access to the project, and wasn't able to figure out how to get one with the information provided.

I did my best to complete the rest of the requirements without being able to access the API, you can see that I put in place some basic error handling and made it fail gracefully. Also if you attempt to access the internal API endpoint without a query parameter, I log the attempt, and provide a message to the user that there is a missing query string.

To summarize what I included in the module:
 - A configuration form available to admins to manage the settings, which I created fields for a base url, the external endpoint, the authorization header as well as the header value. This can be accessed via the admin/config/tableau-accessibility-settings page, or through contextual links on the block itself.
 - Preconfigured settings for the admin form
 - Preconfigured settings for the block instance to be initiated on module install to show up on all pages
 - A twig template with supporting theme function in the .module file to allow passing additional variables to the template file.
 - I setup permissions to allow user roles access to manage the settings, and using the .install file set initial permissions.

Thank you again for the opportunity to show you what I can bring to the company!